refactor: move AIModelInfo in provider_ai_model to remove circular import
